+++ modules/common/src/main/java/edu/tigers/sumatra/drawable/animated/INumberAnimator.java
@@ -14,5 +14,5 @@ public interface INumberAnimator
 	 * 
 	 * @return
 	 */
-	float getNumber();
+	double getNumber();
 }
